requirements.txt
discord.py
aiosqlite
easygoogletranslate
googletrans==4.0.0-rc1
